fix compilation gcc 4.9.3 & 5.2.0

+++ apps/image2cpp.cpp
@@ -121,7 +121,7 @@ int main(int argc, char *argv[])
                                 os_source << tab << tab;
                                 for (coord_t c = 0; c < buff.cols(); c ++)
                                 {
-                                        os_source << std::to_string(buff(r, c));
+                                        os_source << std::to_string(static_cast<unsigned int>(buff(r, c)));
                                         if ((c + 1 < buff.cols()) || (r + 1 < buff.rows()))
                                         {
                                                 os_source << ", ";
